chore(frontend): collapse auth pages into one
